requirements.txt
discord~=0.0.2
unittest2~=1.1.0
web3~=4.8.1

# required by discord... should look into this
discord.py~=0.16.12

# used by !holders command
matplotlib~=3.0.1
numpy~=1.15.3
beautifulsoup4~=4.6.3

# required by web3 AND explicitly imported by this app
requests~=2.20.0

# required by discord.py AND explicitly imported by this app
websockets~=6.0
aiohttp~=1.0.5

# used by websocket.. shouldn't need to explicitly install
# backports.functools_lru_cache~=1.5
# wsaccel~=0.6.2